+++ packages/polaris-viz/src/components/ChartErrorBoundary/tests/ChartErrorBoundary.test.tsx
@@ -1,10 +1,13 @@
+/* eslint-disable no-console */
 import React from 'react';
+import {DataGroup, DataSeries} from '@shopify/polaris-viz-core';
 
 import {ChartSkeleton} from '../../ChartSkeleton';
 import {ChartErrorBoundary} from '../ChartErrorBoundary';
 import {mountWithProvider} from '../../../test-utilities/mountWithProvider';
 
 const MOCK_PROPS = {
+  data: [],
   dimensions: {
     width: 900,
     height: 400,
@@ -15,19 +18,22 @@ function Child() {
   throw new Error();
 }
 
+const warnMock = jest.spyOn(console, 'warn').mockImplementation();
+
 describe('<ChartErrorBoundary />', () => {
-  // eslint-disable-next-line no-console
   const realError = console.error;
 
   beforeAll(() => {
     // Remove the error message from the console because
     // we're expected to have an error.
-    // eslint-disable-next-line no-console
     console.error = (...x) => {};
   });
 
+  afterEach(() => {
+    warnMock.mockReset();
+  });
+
   afterAll(() => {
-    // eslint-disable-next-line no-console
     console.error = realError;
   });
 
@@ -48,6 +54,81 @@ describe('<ChartErrorBoundary />', () => {
         </ChartErrorBoundary>,
       );
       expect(wrapper.text()).toBe('Visible!');
+    });
+  });
+
+  describe('checkForMismatchedData()', () => {
+    it('displays a warning when DataSeries is mismatched', () => {
+      const badData: DataSeries[] = [
+        {
+          name: '',
+          data: [
+            {key: '', value: 1},
+            {key: '', value: 1},
+          ],
+        },
+        {
+          name: '',
+          data: [{key: '', value: 2}],
+        },
+      ];
+
+      mountWithProvider(
+        <ChartErrorBoundary type="Default" {...MOCK_PROPS} data={badData}>
+          <Child />
+        </ChartErrorBoundary>,
+      );
+
+      expect(warnMock.mock.calls[0][0]).toStrictEqual(
+        'The DataSeries[] provided does not have equal series values.',
+      );
+    });
+
+    it('displays a warning when DataGroup is mismatched', () => {
+      const badData: DataGroup[] = [
+        {
+          shape: 'Line',
+          series: [
+            {
+              name: '',
+              data: [
+                {key: '', value: 1},
+                {key: '', value: 1},
+              ],
+            },
+            {
+              name: '',
+              data: [{key: '', value: 2}],
+            },
+          ],
+        },
+        {
+          shape: 'Bar',
+          series: [
+            {
+              name: '',
+              data: [
+                {key: '', value: 1},
+                {key: '', value: 1},
+              ],
+            },
+            {
+              name: '',
+              data: [{key: '', value: 2}],
+            },
+          ],
+        },
+      ];
+
+      mountWithProvider(
+        <ChartErrorBoundary type="Default" {...MOCK_PROPS} data={badData}>
+          <Child />
+        </ChartErrorBoundary>,
+      );
+
+      expect(warnMock.mock.calls[0][0]).toStrictEqual(
+        'The DataGroup[] provided does not have equal series values.',
+      );
     });
   });
 });