+++ packages/polaris-viz/src/components/HorizontalBarChart/tests/Chart.test.tsx
@@ -17,6 +17,7 @@ import {
   VerticalGridLines,
 } from '../components';
 import {normalizeData} from '../../../utilities';
+import {TooltipWrapper} from '../../TooltipWrapper';
 
 jest.mock('../../Labels/utilities/getWidestLabel', () => {
   return {
@@ -306,6 +307,22 @@ describe('<Chart />', () => {
 
       expect(chart).toContainReactComponent(HorizontalBarChartYAnnotations);
       expect(chart).toContainReactComponent(HorizontalBarChartXAnnotations);
+    });
+  });
+
+  describe('<TooltipWrapper />', () => {
+    it('does not render <TooltipWrapper /> data series is empty', () => {
+      const chart = mount(
+        <Chart {...MOCK_PROPS} data={[{name: 'Empty', data: []}]} />,
+      );
+
+      expect(chart).not.toContainReactComponent(TooltipWrapper);
+    });
+
+    it('renders <TooltipWrapper /> data series has data', () => {
+      const chart = mount(<Chart {...MOCK_PROPS} />);
+
+      expect(chart).toContainReactComponent(TooltipWrapper);
     });
   });
 });